fixed WatchedCounter ctor param type

+++ src/engine/mod_bind_benchmark.cpp
@@ -28,7 +28,7 @@ struct WatchedCounter
     /*
      * Exit program after num reports.
      */
-    WatchedCounter(unsigned num = 1) : exit_count(num << shift) { t0 = clock(); }
+    WatchedCounter(size_t num = 1) : exit_count(num << shift) { t0 = clock(); }
     void tick()
     {
         if (bool(++counter & mask) != last_masked)
